added a String extension to simplfy string entry in functions
