--- MarsMission/Helper/Constants.swift
@@ -1,33 +1,0 @@
-//
-//  Constants.swift
-//  MarsMission
-//
-//  Created by Thanyani on 2020/11/03.
-//
-
-import Foundation
-
-
-struct Keys {
-	struct NibName {
-		static let forecastViewController = "ForecastViewController"
-		static let forecastDetailViewController = "ForecastDetailViewController"
-	}
-	
-	struct NavigationTitle {
-		static let forecastCollectionViewControllerTitle = "Mars Mission"
-		static let forecastDetailViewControllerTitle = "Mars Mission"
-	}
-	
-	struct Endpoint {
-		static let url = "https://run.mocky.io/v3/04dc1be1-8609-48c9-b4a0-27a363aa22a9"
-	}
-	
-	struct Identifier {
-		static let reuseIdentifier = "Cell"
-	}
-	
-	struct ForecastTitle {
-		static let title = "Mars Weather Forecast"
-	}
-}
