+++ MarsMission/Helper/Extension.swift
@@ -1,0 +1,23 @@
+//
+//  Extension.swift
+//  MarsMission
+//
+//  Created by Thanyani on 2020/11/03.
+//
+
+import Foundation
+import UIKit
+
+extension String {
+	static let forecastUrl = "https://run.mocky.io/v3/04dc1be1-8609-48c9-b4a0-27a363aa22a9"
+	static let forecastViewController = "ForecastViewController"
+	static let forecastDetailViewController = "ForecastDetailViewController"
+	static let forecastCollectionViewControllerTitle = "Mars Mission"
+	static let forecastDetailViewControllerTitle = "Mars Mission"
+	static let reuseIdentifier = "Cell"
+	static let forecastTitle = "Mars Weather Forecast"
+}
+
+struct ForecastTitle {
+	static let title = "Mars Weather Forecast"
+}
